% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomicParams-class.R
\docType{class}
\name{BioeconomicParams-class}
\alias{BioeconomicParams-class}
\title{A class for bioeconomic parameters.}
\description{
A class for bioeconomic parameters.
}
\section{Slots}{

\describe{
\item{\code{g}}{The magnitude of commercial trip change per year.}

\item{\code{p}}{The magnitude of recreational trip change per year.}

\item{\code{b}}{Recreational catch per effort threshold.}

\item{\code{qC}}{Commercial catchability per effort.}

\item{\code{qR}}{Recreational catchability per effort.}

\item{\code{muA}}{The natural mortality of the adult stock.}

\item{\code{muJ}}{The natural mortality of the juvenile stock.}

\item{\code{chiR}}{Effort per recreational trip.}

\item{\code{chiC}}{Days per commercial trip.}

\item{\code{theta}}{Max number of commercial trips per year.}

\item{\code{gamma}}{Max number of recreational trips per year.}

\item{\code{delta}}{Proportion of recreational catch released.}

\item{\code{varphi}}{Proportion of released recreational catch that survives.}

\item{\code{eta}}{Commercial discard rate.}

\item{\code{Gamma}}{Proportion of commercial discards that survive.}

\item{\code{WR}}{Average weight of recreationally landed fish (tonnes).}

\item{\code{WC}}{Average weight of commercial landed fish (tonnes).}

\item{\code{nu}}{Commercial fish price per tonnes.}

\item{\code{baromega}}{Variable cost per commercial day at sea.}

\item{\code{Lambda}}{Proportion fixed cost attributable to sea bass.}

\item{\code{phi}}{Fixed cost per fleet per year.}

\item{\code{sigma}}{GVA multiplier for commercial Revenue.}

\item{\code{lambda}}{Average per trip expenditure of recreational anglers.}

\item{\code{zeta}}{GVA multiplier for recreational trip expenditure.}

\item{\code{initialSJ}}{The starting value of the size of the juvenile stock (in numbers).}

\item{\code{initialSA}}{The starting value of the size of the adult stock (in numbers).}

\item{\code{initialTC}}{The starting value for the number of commercial trips.}

\item{\code{initialTR}}{The starting value for the number of recreational trips.}
}}

