% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishRAM-package.R
\docType{package}
\name{FishRAM-package}
\alias{FishRAM-package}
\alias{FishRAM}
\title{Fisheries Resource Allocation Model}
\description{
The \code{FishRAM} package implements the dynamic system model from Tidbury et. al. (2021) which simulates the biological and economic impacts for recreational and commercial fishing under different stock allocation scenarios. Default values are set to represent the biology of Sea bass \emph{Dicentrarchus labrax} in division  27-4.b–c (North Atlantic stock) and English inshore commercial fishing fleet and recreational anglers.
}
\details{
Running simulations in \code{FishRAM} is done in 3 main stages:
\enumerate{
\item Set up model parameters using the \code{BioeconomicParams} constructor function.
\item Run simulations using the \code{project} function and a recruitment scenario. This creates a \code{BioeconomicSim} object.
\item Explore results by calling plotting functions (\verb{plot, plotStock, plotActivity, plotEconomicImpact}) on the \code{BioeconomicSim} object.
}
}
\references{
Hannah J Tidbury, Angela Muench, Philip D Lamb, Kieran Hyder, "Balancing biological and economic goals in commercial and recreational fisheries: systems modelling of sea bass fisheries", \emph{ICES Journal of Marine Science}, Volume 78, Issue 5, August 2021, Pages 1793-1803
}
