% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params}
\alias{params}
\title{Sea Bass}
\format{
An object of class \code{"BioeconomicParams"}.
}
\usage{
data("seabass")
}
\description{
A dataset containing the fitted parameters from Tidbury et. al. (2021). These values are designed to represent the biology of Sea bass \emph{Dicentrarchus labrax} in division  27-4.b–c (North Atlantic stock) and English inshore commercial fishing fleet and recreational anglers.
}
\examples{
data("seabass")
sim <- project(params, R = 20, t_end = 40)
plot(sim)
}
\references{
Hannah J Tidbury, Angela Muench, Philip D Lamb, Kieran Hyder, Balancing biological and economic goals in commercial and recreational fisheries: systems modelling of sea bass fisheries, ICES Journal of Marine Science, Volume 78, Issue 5, August 2021, Pages 1793–1803,
}
\keyword{datasets}
