% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomicSim-class.R
\name{plot.BioeconomicSim}
\alias{plot.BioeconomicSim}
\title{Summary plot for \code{BioeconomicSim} objects.}
\usage{
\method{plot}{BioeconomicSim}(x, ...)
}
\arguments{
\item{x}{A \code{BioEconomicSim} object}

\item{...}{Additional plotting parameters. Currently unused.}
}
\description{
After running a simulation, produces 6 plots: stock size, activity (commercial fleet), activity (recreational fleet), economic impact (total across both fleets), economic impact (commercial fleet), economic impact (recreational fleet).
}
\examples{
#loads in the relevant parameters
data("seabass")
# Runs the simulation with 20 years of high recruitment followed by 20 years of low recruitment
R <- c(rep(23151200, 20), rep(757600, 20))
sim <- project(params, R = R)
plot(sim)
}
