% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotActivity}
\alias{plotActivity}
\title{Plot the activity (number of trips) for a simulation.
Under the assumption that recreational and commercial fishers usually conduct daytrips, this indicator can be used as an effort measure.}
\usage{
plotActivity(sim, fleet = "B")
}
\arguments{
\item{sim}{A \code{BioEconomicSim} object.}

\item{fleet}{A \code{character} specifying which fleet to plot. This should be one of \code{"C"} to plot the commercial fleet, \code{"R"} to plot the recreational fleet, or \code{"B"} to plot both fleets on the same axis.}
}
\description{
Plot the activity (number of trips) for a simulation.
Under the assumption that recreational and commercial fishers usually conduct daytrips, this indicator can be used as an effort measure.
}
\examples{
#load in the model parameters
data("seabass")
R <- c(rep(23151200, 20), rep(757600, 20))
sim <- project(params, R = R)
plotActivity(sim)
plotActivity(sim, fleet = "R")
plotActivity(sim, fleet = "C")
}
