% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotEconomicImpact}
\alias{plotEconomicImpact}
\title{Plot the economic impact for a simulation}
\usage{
plotEconomicImpact(sim, fleet = "B")
}
\arguments{
\item{sim}{A \code{BioEconomicSim} object.}

\item{fleet}{A \code{character} specifying which fleet to plot. This should be one of \code{"C"} to plot the commercial fleet, \code{"R"} to plot the recreational fleet, "T" to plot the total GVA, or \code{"B"} to plot both fleets on the same axis.}
}
\description{
Plot the economic impact for a simulation
}
\examples{
#load in the model parameters
data("seabass")
R <- c(rep(23151200, 20), rep(757600, 20))
sim <- project(params, R = R)
plotEconomicImpact(sim)
plotEconomicImpact(sim, fleet = "T")
plotEconomicImpact(sim, fleet = "C")

}
