% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{project}
\alias{project}
\title{Project a bioeconomic simulation}
\usage{
project(
  params,
  R,
  t_start = 1,
  t_end = (length(R) + t_start - 1),
  R_init = NULL
)
}
\arguments{
\item{params}{An object of type \code{BioeconomicParams} encoding the model parameters.}

\item{R}{The recruitment of the stock across the simulation. See details}

\item{t_start}{The year which forms the start of the simulation. The default is 1.}

\item{t_end}{The year which forms the end of the simulation. The default is \code{length(R) + t_start - 1}}

\item{R_init}{The initial recruitment value. Only required if \code{R} is a function of stock size.}
}
\description{
Project a bioeconomic simulation
}
\details{
Recruitment can be specified in 3 ways using different forms of the \code{R} parameter:
\itemize{
\item A \code{numeric} of length 1, giving the constant recruitment across the entire duration of the simulation.
\item A \code{numeric} of length \code{t_end - t_start + 1}, specifying the recruitment at each time step of the simulation.
\item A \code{function} which calculates recruitment from the adult stock size at each time step. This function should have exactly one argument (the adult stock size) and return a \code{numeric} of length 1.
}
}
\examples{
#loads in the relevant parameters
data("seabass")
sim <- project(params, R = 8606000, t_start = 1990, t_end = 2030)

# Runs the simulation with 20 years of high recruitment followed by 20 years of low recruitment
R <- c(rep(23151200, 20), rep(757600, 20))
sim <- project(params, R = R)

#Generates recruitment from a type II functional response recruitment function with log Gaussian noise.
rec_func <- function(stock){
  return(
    1.5@*stock/(7e-4 + 1.3e-4 @* 1.5e-3 @* stock)@*exp(rnorm(1,mean=0,sd=0.9))
  )
}
sim <- project(params, R = rec_func, t_start = 2010, t_end = 2050, R_init = 1e4)
}
