% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomicSim-class.R
\docType{class}
\name{BioeconomicSim-class}
\alias{BioeconomicSim-class}
\title{A class to hold the results of a simulation.}
\description{
A class to hold the results of a simulation.
}
\details{
BioeconomicSim
}
\section{Slots}{

\describe{
\item{\code{params}}{A \code{BioeconomicParams} object encoding the parameters for the model.}

\item{\code{rec_func}}{The recruitment function}

\item{\code{t_start}}{The start time of the simulation.}

\item{\code{t_end}}{The end time of the simulation.}

\item{\code{states}}{A \code{data.frame} containing the states of the system for each time step of the simulation.}
}}

\seealso{
\link{BioeconomicSim}
}
