% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioeconomicSim-class.R
\name{BioeconomicSim}
\alias{BioeconomicSim}
\title{Constructor for the \code{BioeconomicSim} class}
\usage{
BioeconomicSim(params, R, t_start, t_end, R_init = NULL)
}
\arguments{
\item{params}{An object of class \code{BioeconomicParams}.}

\item{R}{The recruitment across the simulation. This is either a \code{numeric} giving the recruitment at each time step, or a function of the previous years adult stock abundance.}

\item{t_start}{An \code{integer} specifying the start time for the simulation. The default is \code{1}}

\item{t_end}{An \code{integer} specifying the end time for the simulation. The default is \code{1 + length(R)}}

\item{R_init}{A \code{numeric} giving the initial recruitment value. Only required if \code{R} is a function of stock size.}
}
\value{
An object of type \linkS4class{BioeconomicSim}
}
\description{
A constructor for the \code{BioeconomicSim} class. This need not be called directly, as this class is generated automatically via the \code{project()} function.
}
\examples{
data("seabass")
recruitment_fun <- function(stock){
 return(
       1.492*stock/(7.2e-4 + 1.87e-07*stock)*exp(rnorm(1,mean=0,sd=0.9))
       )
}
sim <- BioeconomicSim(params, recruitment_fun, t_start = 1980, t_end = 2020, R_init = 1e5)
}
