% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishRAM-package.R
\docType{package}
\name{FishRAM-package}
\alias{FishRAM-package}
\alias{FishRAM}
\title{Fisheries Resource Allocation Model}
\description{
The \code{FishRAM} package implements the model from Tidbury et. al. (2021) to simulate biological and economic impacts of recreational and commercial fleets.
}
\details{
Running simulations in \code{FishRAM} is done in 3 main stages:
\enumerate{
\item Set up model parameters using the \code{BioeconomicParams} constructor function.
\item Run simulations using the \code{project} function and a recruitment scenario. This creates a \code{BioeconomicSim} object.
\item Explore results by calling plotting functions (\verb{plot, plotStock, plotActivity, plotEconomicImpact}) on the \code{BioeconomicSim} object.
}
}
\references{
Hannah J Tidbury, Angela Muench, Philip D Lamb, Kieran Hyder, "Balancing biological and economic goals in commercial and recreational fisheries: systems modelling of sea bass fisheries", \emph{ICES Journal of Marine Science}, Volume 78, Issue 5, August 2021, Pages 1793-1803
}
